% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_eastneighbour.R
\name{pluscode_eastneighbour}
\alias{pluscode_eastneighbour}
\title{Pluscode East Neighbour}
\usage{
pluscode_eastneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the easterly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_eastneighbour("9C4MGC2M+H4")

}

