% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_neighbours.R
\name{pluscode_neighbours}
\alias{pluscode_neighbours}
\title{Pluscode Neighbours}
\usage{
pluscode_neighbours(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This function retrieves the neighbouring pluscodes of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_neighbours("9C4MGC2M+H4")

}

