% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_northeastneighbour.R
\name{pluscode_northeastneighbour}
\alias{pluscode_northeastneighbour}
\title{Pluscode Northeast Neighbour}
\usage{
pluscode_northeastneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the north-easterly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_northeastneighbour("9C4MGC2M+H4")

}

