% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_northneighbour.R
\name{pluscode_northneighbour}
\alias{pluscode_northneighbour}
\title{Pluscode North Neighbour}
\usage{
pluscode_northneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the northerly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_northneighbour("9C4MGC2M+H4")

}

