% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_northwestneighbour.R
\name{pluscode_northwestneighbour}
\alias{pluscode_northwestneighbour}
\title{Pluscode Northwest Neighbour}
\usage{
pluscode_northwestneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the north-westerly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_northwestneighbour("9C4MGC2M+H4")

}

