% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_southeastneighbour.R
\name{pluscode_southeastneighbour}
\alias{pluscode_southeastneighbour}
\title{Pluscode Southeast Neighbour}
\usage{
pluscode_southeastneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the south-easterly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_southeastneighbour("9C4MGC2M+H4")

}

