% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_southneighbour.R
\name{pluscode_southneighbour}
\alias{pluscode_southneighbour}
\title{Pluscode South Neighbour}
\usage{
pluscode_southneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the southerly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_southneighbour("9C4MGC2M+H4")

}

