% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscode_westneighbour.R
\name{pluscode_westneighbour}
\alias{pluscode_westneighbour}
\title{Pluscode West Neighbour}
\usage{
pluscode_westneighbour(pluscode)
}
\arguments{
\item{pluscode}{A valid pluscod 2, 4, 8, or 10 characters in length excluding the plus sign (which can be included)}
}
\description{
This package retrieves the westerly neighbour of a valid pluscode, with a precision of 2, 4, 8, or 10 excluding the plus sign (which can be included)
}
\examples{
pluscode_westneighbour("9C4MGC2M+H4")

}

